requirements.txt
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.0.0
python-dotenv>=1.0.1
python-docx>=1.1.0
openai>=1.40.0
httpx>=0.27.0
